dalsze prace

+++ src/main/java/pl/wsikora/successbudget/v3/common/type/application/MoneyDto.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Value
 public class MoneyDto {
 
-    Long currencyId;
+    Integer currencyId;
     BigDecimal value;
 
 }
